+++ spec/directions/bake_numbered_exercise/v1_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-RSpec.describe Kitchen::Directions::BakeNumberedExercise::V1 do
+RSpec.describe Kitchen::Directions::BakeNumberedExercise do
   let(:exercise1) do
     book_containing(html:
       one_chapter_with_one_page_containing(
@@ -22,7 +22,7 @@ RSpec.describe Kitchen::Directions::BakeNumberedExercise::V1 do
 
   context 'when solutions are not suppressed' do
     it 'works' do
-      described_class.new.bake(exercise: exercise1, number: '1.1')
+      described_class.v1(exercise: exercise1, number: '1.1')
 
       expect(exercise1).to match_normalized_html(
         <<~HTML
@@ -48,7 +48,7 @@ RSpec.describe Kitchen::Directions::BakeNumberedExercise::V1 do
 
   context 'when solutions are suppressed' do
     it 'works' do
-      described_class.new.bake(exercise: exercise1, number: '1.1', suppress_solution_if: true)
+      described_class.v1(exercise: exercise1, number: '1.1', suppress_solution_if: true)
 
       expect(exercise1).to match_normalized_html(
         <<~HTML
@@ -69,7 +69,7 @@ RSpec.describe Kitchen::Directions::BakeNumberedExercise::V1 do
   context 'when even solutions are suppressed' do
     context 'when number is odd' do
       it 'works' do
-        described_class.new.bake(exercise: exercise1, number: 1, suppress_solution_if: :even?, note_suppressed_solutions: true)
+        described_class.v1(exercise: exercise1, number: 1, suppress_solution_if: :even?, note_suppressed_solutions: true)
         expect(exercise1).to match_normalized_html(
           <<~HTML
             <div data-type="exercise" id="exercise_id" class="os-hasSolution">
@@ -94,7 +94,7 @@ RSpec.describe Kitchen::Directions::BakeNumberedExercise::V1 do
 
     context 'when number is even' do
       it 'works' do
-        described_class.new.bake(exercise: exercise1, number: 2, suppress_solution_if: :even?, note_suppressed_solutions: true)
+        described_class.v1(exercise: exercise1, number: 2, suppress_solution_if: :even?, note_suppressed_solutions: true)
         expect(exercise1).to match_normalized_html(
           <<~HTML
             <div data-type="exercise" id="exercise_id" class="os-hasSolution-trashed">
@@ -110,5 +110,80 @@ RSpec.describe Kitchen::Directions::BakeNumberedExercise::V1 do
         )
       end
     end
+
+    context 'when there is a stimulus' do
+      let(:multipart_exercise) do
+        book_containing(html:
+          one_chapter_with_one_page_containing(
+            <<~HTML
+              <div data-type="exercise" id="ex1">
+                <div data-type="problem" id="prob1">
+                  <div class="question-stimulus">Stimulus</div>
+                  <div class="question-stem">question 1</div>
+                  <div class="question-stem">question 2</div>
+                  <div class="question-stem">question 3</div>
+                </div>
+              </div>
+            HTML
+          )
+        ).chapters.exercises.first
+      end
+
+      let(:multipart_exercise_with_one_part) do
+        book_containing(html:
+          one_chapter_with_one_page_containing(
+            <<~HTML
+              <div data-type="exercise" id="ex1">
+                <div data-type="problem" id="prob1">
+                  <div class="question-stimulus">Stimulus</div>
+                  <div class="question-stem">question 1</div>
+                </div>
+              </div>
+            HTML
+          )
+        ).chapters.exercises.first
+      end
+
+      it 'bakes a multipart exercise' do
+        described_class.v1(exercise: multipart_exercise, number: 2)
+        expect(multipart_exercise).to match_normalized_html(
+          <<~HTML
+            <div data-type="exercise" id="ex1">
+              <div data-type="problem" id="prob1">
+                <span class="os-number">2</span>
+                <span class="os-divider">. </span>
+                <div class="os-problem-container">
+                  <div class="question-stimulus">Stimulus</div>
+                  <ol type="a">
+                    <li><div class="question-stem">question 1</div></li>
+                    <li><div class="question-stem">question 2</div></li>
+                    <li><div class="question-stem">question 3</div></li>
+                  </ol>
+                </div>
+              </div>
+            </div>
+          HTML
+        )
+      end
+
+      it 'doesn\'t add a list when there is just one part' do
+        described_class.v1(exercise: multipart_exercise_with_one_part, number: 2)
+        expect(multipart_exercise_with_one_part).to match_normalized_html(
+          <<~HTML
+            <div data-type="exercise" id="ex1">
+              <div data-type="problem" id="prob1">
+                <span class="os-number">2</span>
+                <span class="os-divider">. </span>
+                <div class="os-problem-container">
+                  <div class="question-stimulus">Stimulus</div>
+                  <div class="question-stem">question 1</div>
+                </div>
+              </div>
+            </div>
+          HTML
+        )
+      end
+    end
+
   end
 end